[test] (ABC-662): did two things: - separated tool managers in files and - added clear undo/redo test (green coverage for all except smooth tools)
