+++ src/modules/base/inputPen/inputPen.html
@@ -246,10 +246,8 @@
                 </div>
                 <div
                     class={computedTextAreaClasses}
-                    onmouseup={handleMouseUp}
-                    onmousedown={handleMouseDown}
-                    onmousemove={handleMouseMove}
                     onmouseenter={handleMouseEnter}
+                    onmousedown={handleMouseDown}
                     onmouseleave={handleMouseLeave}
                     data-element-id="draw-area"
                 >
